feat(runner): add chain runner
Add command runner which works in chain mode. This library will
be used by the executable, which will handle command line
processing and JSON parsing and serialization.

In chain mode, the command is prefixed with a Lodash chain with
the data as the context. After the command, .value() is
automatically called.