style(run): fix uncaught spacing errors
The jscs config requires spaces between curly braces and their contents.

However, an issue with the gulp config means that files are sometimes
not checked with jscs. !1

Fortunately, jscs seems to correctly run in Travis (but only as of
recent commits, Travis was happy with these lines when they were
initially commited).