fix(run): remove explicit calling of chain value()
Lodash has a toJSON method on chain objects. This means that, when
a chain, or any data structure containing a chain, is passed to
JSON.stringify, value() is implicitly called everywhere necessary.

A side effect of this change is that we must now explicitly call
value() in unit tests of run.run, since value() is only called when
stringifying.