style(airbnb): add space before function parens
Enforce #7.11 from the Airbnb style guide, a function's opening
parenthesis must have a space before it if there is no function identifier.
https://github.com/airbnb/javascript/blob/c25dbac620b258c4421251bc403fffa1051de61e/README.md#7.11

Originally, jowl had a transitive dependency upon jscs 2.7.0, which did
not enforce this rule. However, jscs 2.9.0 and later do enforce it, and
those who installed jowl after it came out picked up that version of jscs
because of the semver ranges specified for gulp-jscs by jowl and jscs by
gulp-jscs.

Fortunately, this was easy to fix automatically by running `jscs -x` on
this project.

Thanks to hbetts for reporting this in #2 .