feat(run): limit p() to taking one argument
Limit `p()` to taking only one argument, even though `console.json()`
can handle an unlimited number of arguments.

This is an ergonomic improvement. It is natural to want to pass `p()`
as an iteratee to lodash, and lodash passes a number of additional
optional arguments (that users often forget the existence of) to the
iteratee as well.

Concretely, this is the difference between

```
echo '[1,2,3]' | jowl -q 'c.each(function(a){p(a)})'
```

and

```
echo '[1,2,3]' | jowl -q 'c.each(p)'
```

This does mean an extra step is needed to combine keys and values
when iterating over an object, but the user would need to do that
anyway in the likely event that they want a format other than
`value key`.