docs(reference): mind our p()'s and -q's
Document the `p` function and `--quiet` option.

`p()` has its own example rather than being part of the combined
"Variables" example, because the latter was becoming too crowded
and because `p()` already has its own subsection.

Move the "Input and Output" section higher up, it is necessary to
understanding jowl's functionality, even without any of the
customizations in the "Variables" and "Options" sections.