chore(markdown): add markdownlint
Add linting for Markdown via markdownlint and fix identified errors.

Markdownlint has been invaluable in the past for finding problems
that cause renderers to disagree and making good suggestions for
authoring.

Line length checking has been disabled globally since the ideal way
to write Markdown under source control is one sentence per line.

Specifically disabled rule for dollar sign verbatim text being
followed by output for the README because the dollar sign adds
immediate visual context for the command line, but the top of the
readme needs to be concise and is not the place to add output.