chore(lint): switch to current Airbnb rules
Prepare to move to es6 by switching from Airbnb ES5 rules to their
current ruleset. Apply warnings for all rules we will implement; no
source has changed in this commit.

The trickiest rule to understand was their handling of 'use strict'.
They set `sourceType` to `module`, which led to complaints about its
use in script files (and Node will require it to be enabled for const
and let). Add a second level eslintrc in the lib folder to
differentiate which files are in fact modules.

They also forbid all 'use strict' because their babel settings add it.
Reenable it.