style(lint): use template strings instead of concat
Satisfy `prefer-template` rule by substituting templates wherever
string concatenation was previously used.

While I like this ES6 feature in general, I actually feel like this
made the integration test that appened a newline to a computed string
harder to read. However, I will bow to the wisdom of the rule for the
sake of consistency, and hope that it is not one of those foolish
consistency hobgoblins.