docs(reaadme): add installation instructions
Add instructions on how to install jowl.

`--global` is appropriate because jowl is inteneded to be a command
line tool on peoples' $PATH, and no node project will depend on it.

`--production` is important because jowl has tons of devDependencies
that are completely unecessary for everyday use.

Standardize on fenced code blocks for the readme so
that a language can be specified.

Fix a markdownlint-enable tag.