chore: files are always LF
Attempt to ensure that files always have LF (Unix) line endings, in the
object database as well as when checked out.

This is almost completely enforced with a .gitattributes, except that
if a developer creates a CRLF file, git won't change it to LF in their
working directory after it's committed, only if they delete it and
check it out again. Therefore, also add a .editorconfig and hope that
editor obey it.

I had really hoped to allow people to use platform-native line endings
and convert everything to lf within the object database, but lint rules
and `npm publish` made that impossible.