docs: add windows support
Explicitly note our supported OSes in the README and CONTRIBUTING.

Unfortunately, due to escaping rules, it doesn't seem possible to write
examples that work in both a Bourne shell and CMD or PowerShell. The
best we can do for now is note that the examples are written with
Bourne shell in mind.

Add Appveyor status badge to CONTRIBUTING.