feat(cli): pass through when no command
If the user does not supply a JavaScript command, treat it as if they
had supplied a command of 'd' rather than simply printing `undefined`.
This has the effect of turning jowl with no arguments into a
pretty-printer.

This was inspired by my own failures to use jowl correctly while
testing it on Windows and wondering what terrible shell escaping was
turning all of the stdin expressions into undefined.

Originally, I implemented an error message to let the user know they
needed to supply one, but a default passthrough is much more useful.