docs(contributing): miscelaneous tweaks
Add thanks to the list of reasons that issues are welcomed.

Reword reasoning for filing an issue before a merge request to confirm
that it's not mandatory.

When mentioning CI in the Developing docs, link to both Travis and
Appveyor. This change was missed earlier when Appveyor was added.

Note that it's fine to change docs inline rather than as a separate
change.