chore(deps): switch to yarn
Add a yarn.lock for reproducible installs and change CI to use it.

Add Yarn as a necessary tool for development; but end-users are still
welcome to use whatever package manager they like. Install instructions
are kept in terms of npm so that they can be followed with a plain Node
install, but Yarn users will know what to do.