chore(circle): do not restrict version of Yarn
Stop restricting yarn to a particular version in CI that was only pinned
to fix an incompatability with Node 5, which we no longer support.

We should think about how to manage the Yarn 1 vs Yarn 2
incompatability, but Yarn 1 will continue to be what we get when we
instlal plain `yarn` for quite a while.