chore(lint): bump eslint and airbnb versions
Bump eslint from major version 3 to 5, which is the last version to
support Node 6. This removes "Warning: Accessing non-existent property
of 'foo' of module xports inside circular depndency' warnings (with many
different values of 'foo') during eslint runs on Node 14.

For compatability with eslint, upgrade to the latest version of the
airbnb rules. There is now an eslint-config-airbnb-base, which no longer
requires eslint-plugin-jsx-a11y or eslint-plugin-react as
peerDependencies, so pick the latest version of it and get rid of those.

Since we have gone through several breaking changes to the airbnb eslint
rules, disable new rules that the codebase does not comply with. They
all do make sense to fix, but this should be done at a later date to
avoid conflating Node compatability with code fixes for linting.