style(lint): configure elsint with json
Switch from .eslintrc.js to .eslintrc.json. Our only reason for picking
JavaScript in the first place was that it allowed comments, and eslint
now accepts a superset of JSON that includes comments in .eslintrc.json

Ironically, the .eslintrc.js was written in an almost-JSON already, and
didn't actually follow the rest of our linting rules.

This is a formatting-only change to the .eslintrc, no rules have been
altered.