chore(lint): fix bugs in eslint patterns
Fix several bugs in the patterns passed to the eslint tool as part of
linting this codebase.

First, remove the `*.js` rule. It didn't actually match the .eslintrc.js
because, like a shell glob, files with a leading dot were ignored. Now
that the eslintrc is json, there are no js files in the project root
anyway.

Second, there was a superfluous comma after the rule matching files in
src that came from converting from gulp to a shell command. This
prevented any files in src from matching.