chore(deps): explicitly declare yarn 1 compatability
Set yarn's version in package.json's engines to declare that we want
something compatible with Yarn 1. Yarn 2 has several breaking changes
and I have not yet had time to evaluate them and decide about using Yarn
2. Be clear about this in the CONTRIBUTING as well.

As far as I can tell, this should not prevent npm from installing this
package. We do not set engineStrict, and npm is looking for npm and node
versions in engines, not yarn's version.