chore: fix script globs for windows
Use double-quotes instead of single-quotes to quote globs in all of the
package.json scripts for compatability with the Windows CMD shell.

Globs need to be quoted so that they are not expanded by the shell, but
single quotes are passed by Windows directly to the process, leading
eslint to fail because it cannot find files whose paths start and end
with a single quote. The eslint docs explicitly mention that
double-quotes should be used on Windows:
https://eslint.org/docs/user-guide/command-line-interface

Single quotes remain fine for Unix as well, since we're not using any
shell metacharacters within the globs.