style: comply with operator-linebreak
Reformat files to move the operator to the beginning of a line rather
than the end of the previous line on multiline statements.

This appears to be the intent of the Airbnb style guide according to
https://github.com/airbnb/javascript/pull/1600 even though it is not
incredibly clear from the way the guide is written. The sentence in the
guide about this rule, at first reading, appears to only apply to
logical operators within control statement conditions.

Nevertheless, this rule is more readable.