feat(format): indent to 2 spaces instead of 4
Change the default indentation to 2 spaces per level rather than 4
spaces per level.

This matches the behavior of json-colorizer when the library is
serializing JSON itself, and we would like to use it in this mode to
avoid the penalty of serializing JSON only to have json-colorizer
immediately tokenize it again. json-colorizer does have an open pull
request to customize indent level, but it is not yet merged
https://github.com/joeattardi/json-colorizer/pull/20 .

This also matches the default behavior of jq.

BREAKING CHANGE:

While the number of spaces to use for indentation was never explicitly
guaranteed by documentation, it was documented through examples in the
reference. Changing the indentation from 4 spaces to 2 has the
possibility of breaking scripts that do string comparisons against the
output of jowl.