refactor(format): add no-op call to json-colorizer
When formatting result data for output, pass it through json-colorizer
as well. However, disable color output from json-colorizer so that this
remains a no-op refactor.

Continue stringifying the JSON ourselves rather than letting
json-colorizer do it so that of the result data consists of a single
string, json-colorizer will not try to parse that string as JSON. This
adds slightly more overhead, but in practice doesn't much impact
performance.

Rework unit tests to ensure that we are actually calling json-colorizer.
This is is easier to do than checking if output has correctly had ANSI
color codes added to it, because chalk (which json-colorizer uses
internally) applies heuristics to determine whether to output color, and
overriding them becomes a game of changing global state in tests to
influence a dependency-of-a-dependency. Color behavior will be better
tested by integration tests once we are actually outputting color.
Unfortunately, because json-colorizer exports a single function, the
most straightforward way to spy on it via sinon is to make that funciton
a property of the exported format object. The alternative would be yet
another devDependency that rewrites how require() works, but that adds
too much complexity to be worth it.

Depend upon chalk ^2.4.2 in order to switch off color codes. This is a
very old version, but is appropriate for several reasons:
* It matches the major version that json-colorizer uses
* It is the last version of chalk that supports Node 6, which we still
  support.