test(format): do not overfit return values
When unit testing formatOutput and runFormat, assert their return values
in terms of the functions they call, not literal values.

Once we begin adding color to output, it will become infeasable to
anticipate or override all of the behavior of the color library in unit
tests, because this behavior depends upon global state like environment
variables, command-line options, and a global singleton setting, and the
unit test runner itself uses the same library for color output.

When we make assertions in terms of the return values of the functions
used by the function under test, we better test that function as a unit.

Color behavior testing can then be left mostly to integration tests,
which have more control over the process environment.