feat(format): add color syntax highlighting
Use json-colorizer to add ANSI color codes to syntax highlight result
data. The color scheme still needs to be tested on more terminals and
themes, but is at least reasonably readable. It is limited to the
16-color space, so no downsampling should be necessary.

Add support for the --color and --no-color options to override chalk's
heuristics on when to output color codes. Wierdly, since chalk reads
these from ARGV directly, there's no plumbing for this other than
telling Commander to not reject them.

Chalk (actually its dependency supports-color) also supports other
values for --color but leave these undocumented until they've been
given more consideration. Most of them serve to change the bit depth
of the colors output, and this doesn't matter because we are only using
colors from the 16-color space. Ironically, it would have been nice to
support the same interface as ls and grep and have --color take the
values "always", "auto", or "never", but this is only supported in a
much later version of supports-color which is only used by a later
version of chalk which doesn't support Node 6 and is also not used by
any version of json-colorize yet. It may eventually be worthwhile
to do our own option handling for --color later, never let chalk read
it, and pass through a value for chalk.level that we use.

Commander doesn't give us a great way yet to declare --color and
--no-color being mutually exclusive, so just document the use of both of
them being undefined behavior for now so that we can add that check
later without it being a breaking change.

Also document FORCE_COLOR=1 and FORCE_COLOR=0 which are also read
directly by chalk. Don't document any other levels for the same color
depth reasons as above.

Add the ability for integration tests to set environment variables to
handle testing FORCE_COLOR. Pass through a few important variables like
PATH by default to make sure that the tests will actually run.