chore(ci): move Windows builds to CircleCI
Consolidate all builds into one CircleCI matrix definition and remove
AppVeyor for Windows builds. This removes definition redundancy and
makes the build process easier to explain.

Add a machine definition for a Windows executor and specify it as part
of the build matrix. While Circle's docs say to use the Windows orb's
executor definition, the syntax for this seems incompatible with the
top-level executor defintions that are necessary for feeding into the
mtarix as parameters. Instead, we expand them to the definition
specififed by the orb.

Circle's Node orb is also incompatible with Windows, but the Windows
executor already has NVM. Conditionally either call NVM or
use the Node orb to set up Node and Yarn. Other orb commands have just
been replaced with their plain equivalents, which turn out to be exactly
the same on all three platforms.

Caching is a bit weird because we need to parameterize the cache on OS,
(filesystem permissions mean that we can't mount a Linux cache on
Windows, for example), but the pipeline only gets an abstract executor object.
We let Circle stringify the executor object to generate the cache key, and this
surprisingly seems to be consistent between runs despite not really
being documented. If it ever changes, the worst case is that we'll get a
cache miss.

Delete the AppVeyor config and remove references to AppVeyor from the
docs. AppVeyor has served us extremely well, especially when they were
the only Windows game in town, but it's time to say goodbye.