chore(ci): add automatic brew tap bumps
Add new deploy-homebrew job to bump the version and checksum in the brew
tap and push the new commit automatically after publishing to npmjs.org.

The closest brew command to do this is "bump-formula-pr", intended to
open a pull request with the formula update. We instead invoke it but
tell it just to make the formula changes and commit them to homebrew's
checked out working directory for the tap. We pass the version (derived
from the tag) explicitly, both to avoid race conditions with the
registry, and also because "bump-formula-pr" can't figure out the new
version even though "bump" can.