chore(ci): add Fedora rpm specfile
Add specfile to build Jowl as an RPM on Fedora. This is largely based
upon the example specfile provided in
https://docs.fedoraproject.org/en-US/packaging-guidelines/Node.js/

Since this is an executable that happens to be written in Node, rather
than a npm library, we can skip several steps that are only necessary
for libraries.

We use a tarball from GitHub as Source0 rather than the npmjs.org
published package (the latter is recommended by the packaging
guidelines) because this project considers the npmjs.org package as a
build product, not the preferred form of the source code. In addition,
the RPM needs tests and docs that aren't part of the npmjs.org package.

Manually fix shebang lines. Normally brp-mangle-shebangs should take
care of this, but it has a bug on fc35 and older where it doesn't find
node executables correctly.