chore(ci): automate Fedora builds with Packit
Add packit config to allow automatic publishing of jowl into Fedora
packaging infrastructure.

The most awkward part of this is handling generation of the bundled node
dependency tarballs and bundled license file. Add new script
dist/packit-fix-spec-file.sh to replace packit's fix-spec-file action.

This script runs nodejs-packaging-bundler to generate the extra files.
Unfortunately that script is hardcoded to dump its output to
rpmbuild/SOURCES, so we must move it back to the dist directory where
packit expects it.

Because the tarball should be identical to one exported from GitHub, the
script does not mangle Source0 the way that Packit normally would. This
provides better documentation of the provenance of the code.