chore(ci): fix specfile to comply with guidelines
Fix specfile based upon review to comply with Fedora Packaging
Guidelines.

Remove npm_name macro. There's no need for an npm_name macro when we can
just use the name macro, since the name of the project on NPM and in
Fedora will always be the same.

Fix dangling node_modules_prod/.bin symlink. The example spec for tape
assumes that the prod node modules tarball will have a .bin directory
in it. Since Jowl's doesn't, don't try to create a symlink to that
directory, and create the directory during check so that we can combine
it with the dev node modules' .bin directory to run tests.

Use Fedora Packaging Gudidelines on Git Hosting Services to choose a
GitHub URL format for the Source0 URL. This URL is presumably more
stable, and is at least more consistent with other projects.

Add a trailing slash to file jowl directory in the files list to ensure
it is installed correctly.

Use fdupes RPM macro to replace duplicate files in node_modules with
links instead.

Add comments about how to generate the remaining sources from the
first one programmatically.

Clarify in comments why jowl/bin/jowl is not a symlink to jowl.js
(doing so messes up Node's ability to resolve the node_modules directory
correctly).