chore(ci): clarify how to build outside of packit
It should be possible for someone to build jowl from dist-git with or
without running packit.

Ensure the packit-fix-spec-file.sh is synced to dist-git so that it can
clearly be seen by maintainers.

More clearly document that the only necessary step of
packit-fix-spec-file is running nodejs-packaging-bundler, so that it can
be easily run outside of packit.