chore(ci): build for Rawhide as well
Copy in the default jobs that were previously implicitly being provided
by packit. Switch their targets to fedora-all in order to build against
all released Fedora versions plus Rawhide.

This package will be initially proposed for Rawhide, so it needs to
build there. In the future, we may reconsider always building against
rawhide as that may cause builds to fail for reasons outside of changes
to code.