chore(ci): configure packit to make dist dir
When packit syncs files from this repo to Fedora's dist-git, explicitly
tell it to create the destination dist/ directory if it does not already
exist.

It is an explicit design decision to require mkpath to be specified for
packit to create directories. See https://github.com/packit/packit/issues/1738
for more information.