chore(deps): fix npm dependency bug for srpm build
When building the SRPM, force /usr/bin/node to be installed, which is a
dependency of nodejs-packaging-bundler.

This works around a bug in the packaging of node.
nodejs-packaging-bundler Requires npm, but since
https://src.fedoraproject.org/rpms/nodejs20/c/92480e045911151ae470f279ceef834f0f5b4370
has not yet been released, that can be Provide-ed by a non-default node
version, which won't actually install a /usr/bin/node (see
https://download.copr.fedorainfracloud.org/results/packit/daxelrod-jowl-53/srpm-builds/06245209/builder-live.log.gz
as an example of this, where on fc38, default node is 18, but
nodejs20-npm gets installed).