chore(ci): dep tarballs in post-upstream-clone
Change the mechanism by which the Packit build generates node module
dependency tarballs.

Previously, we ran nodejs-packaging-bundler during fix-spec-file.
However, fix-spec-file is only intended to make a non-release Git
checkout packageable, and therefore doesn't run during the
propose-downstream job.

Instead, move the invocation to post-upstream-clone, which runs during
all actions and is designed to handle source generation.

To invoke nodejs-packaging-bundler in a way that produces the same
output, we nee to pass it a tarball and a version, but Packit doesn't
provide those during post-upstream-clone. Instead, get the version out
of the package.json and use npm to create an npm dist tarball for use by
nodejs-packaging-bundler.