chore(ci): configure packit for propose-downstream
When running propose-downstream, after rsyncing files from the upstream
repo clone to the dist-git repo, Packit expects all source files to be
in the root of the dist-git repo.

To rsync them there correctly, it is necessary to use the src and
dest flags so that the wildcards are interpreted correctly.